Folders created and files relocated
